Create timesheet CSV editor command
This is a command line tool to download and upload timesheet records.

To download records:

> getmytime-edit.py download 2017-01-01 > timesheet.csv

To upload records:

> getmytime-edit.py upload timesheet.csv

Preview changes by using the `--dry-run` flag.

To delete records, prefix the ID field with "-". The record will be
deleted when you upload the timesheet.

To create records, the ID must be 0 or blank. When you upload the
timesheet, the ID field will become populated.

Updating records is not supported! The workaround is to create a new
record and delete the old one. To do this copy & paste a row, set the ID
to 0, then prefix the old record ID with "-" to delete it.

Remember to refresh the timesheet in your editor after uploading!